refactor: rename tabsConfig file
