+++ components/character-tabs/tabsConfig.ts
